바위, 루트, 게시글 인덱스 오류 수정

+++ src/main/java/com/line7studio/boulderside/application/search/dto/SearchItemResponse.java
@@ -23,10 +23,10 @@ public class SearchItemResponse {
     private String city;
     private Level level;
     private String authorName;
-    private Integer viewCount;
-    private Integer commentCount;
-    private Integer likeCount;
-    private Integer climberCount;
+    private Long viewCount;
+    private Long commentCount;
+    private Long likeCount;
+    private Long climberCount;
     private LocalDate meetingDate;
     private LocalDateTime createdAt;
 }
